Intuitively reorganised project structure for readability and easily navigation / maintainability. Refactoring. Integrated persist task inside the final App Run class.
