+++ src/main/java/com/avangarde/parkinglot/IO/read/ReadFromFile.java
@@ -1,4 +1,4 @@
-package com.avangarde.parkinglot.utils;
+package com.avangarde.parkinglot.IO.read;
 
 import com.avangarde.parkinglot.parking.entities.ParkingLot;
 import com.avangarde.parkinglot.vehicle.entities.Vehicle;
@@ -6,7 +6,7 @@ import com.avangarde.parkinglot.vehicle.entities.Vehicle;
 import java.io.File;
 import java.util.List;
 
-public interface IInputFile {
+public interface ReadFromFile {
 
     ParkingLot readParkingLot(String path);
     List<Vehicle> readVehicles(String path);
