+++ src/main/java/com/avangarde/parkinglot/IO/write/VehicleAsString.java
@@ -1,10 +1,10 @@
-package com.avangarde.parkinglot.vehicle.interfaces;
+package com.avangarde.parkinglot.IO.write;
 
 import com.avangarde.parkinglot.vehicle.entities.Vehicle;
 
 import java.util.List;
 
-public interface IVehicleAsString {
+public interface VehicleAsString {
 
     String write(List<Vehicle> vehicles);
 }
