+++ src/main/java/com/avangarde/parkinglot/database/repositories/ParkingFloorRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.avangarde.parkinglot.repositories;
+package com.avangarde.parkinglot.database.repositories;
 
 import com.avangarde.parkinglot.parking.ParkingSpotFactory;
 import com.avangarde.parkinglot.parking.SpotType;
