+++ src/main/java/com/avangarde/parkinglot/parking/FloorBuilder.java
@@ -1,11 +1,11 @@
-package com.avangarde.parkinglot.parking.interfaces;
+package com.avangarde.parkinglot.parking;
 
 import com.avangarde.parkinglot.parking.ParkingSpotLotSize;
 import com.avangarde.parkinglot.parking.entities.ParkingFloor;
 
 import java.util.List;
 
-public interface IFloorBuilder {
+public interface FloorBuilder {
 
     ParkingFloor createParkingFloor(int floorNo, List<ParkingSpotLotSize> parkingSpotLotSizes);
 }
